latest uncleaned code for clinical pipeline
