+++ examples/clinical_pipeline/search.json
@@ -1,0 +1,5 @@
+{
+  "query": {
+    "bool": {}
+  }
+}
