+++ examples/clinical_pipeline/settings.json
@@ -1,0 +1,3 @@
+{
+	"index.mapping.coerce": false
+}
